fix: game is starting up in neovim 0.10.3
